+++ src/react-apps/applications/shared/src/utils/attachments.ts
@@ -1,0 +1,36 @@
+import { IAttachment, IData } from './../types/index';
+
+export const mapInstanceAttachments = (data: IData[]): IAttachment[] => {
+  if (!data) {
+    return [];
+  } else {
+    const tempAttachments: IAttachment[] = [];
+    data.forEach((dataElement: IData) => {
+      if (dataElement.elementType !== 'default' && dataElement.elementType !== 'ref-data-as-pdf') {
+        tempAttachments.push({
+        name: dataElement.fileName,
+        url: dataElement.dataLinks.apps,
+        iconClass: 'reg reg-attachment' });
+      }
+    });
+    return tempAttachments;
+  }
+};
+
+export const getInstancePdf = (data: IData[]): IAttachment => {
+  if (!data) {
+    return null;
+  }
+
+  const pdfElement = data.find((element) => element.elementType === 'ref-data-as-pdf');
+
+  if (!pdfElement) {
+    return null;
+  }
+
+  return {
+    name: pdfElement.fileName,
+    url: pdfElement.dataLinks.apps,
+    iconClass: 'reg reg-attachment',
+  };
+};
